separate UploadProblemForm component
