+++ src/components/Problems/Problems.module.scss.d.ts
@@ -1,7 +1,6 @@
 export type Styles = {
   actions: string
   container: string
-  dropzone: string
   loading: string
   menu: string
   notPublished: string
