+++ src/components/Problems/UploadProblemForm.module.scss
@@ -1,0 +1,21 @@
+.dropzone {
+  cursor: pointer;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 5rem;
+  font-weight: bold;
+  text-transform: uppercase;
+  color: #BBB;
+  background-color: #EEE;
+}
+
+.dropzone:hover {
+  background-color: #DDD;
+}
+
+.bottomAction {
+  margin-top: .5rem;
+  display: flex;
+  justify-content: flex-end;
+}
