+++ src/components/Problems/UploadProblemForm.module.scss.d.ts
@@ -1,0 +1,10 @@
+export type Styles = {
+  bottomAction: string
+  dropzone: string
+}
+
+export type ClassNames = keyof Styles
+
+declare const styles: Styles
+
+export default styles
